firebaseの認証の処理をapp.jsに移す
