--- program/firebase.js
@@ -1,26 +1,0 @@
-// Import the functions you need from the SDKs you need
-// import { initializeApp } from "firebase/app";
-// import { getAuth } from "firebase/auth";
-// import { getFirestore } from "firebase/firestore"
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: process.env.API_KEY,
-  authDomain: process.env.AUTH_DOMAIN,
-  projectId: process.env.PROJECT_ID,
-  storageBucket: process.env.STORAGE_BUCKET,
-  messagingSenderId: process.env.MESSAGING_SENDER_ID,
-  appId: process.env.APP_ID
-};
-
-// Initialize Firebase
-// const firebaseapp = initializeApp(firebaseConfig);
-// const auth = getAuth(firebaseapp);
-// const db = getFirestore(firebaseapp);
-// provider.addScope("repo");
-
-// export {auth, db}
-
-firebase.initializeApp(firebaseConfig);
-
-export const auth = firebase.auth();
